DAOS-647 common: Split out ACL principal functions
This is preliminary reorganization for adding some new
ACL principal-related utility functions.
