+++ utils/run_test.sh
@@ -116,7 +116,8 @@ if [ -d "/mnt/daos" ]; then
     run_test build/src/security/tests/srv_acl_tests
     run_test build/src/common/tests/acl_api_tests
     run_test build/src/common/tests/acl_util_tests
-    run_test build/src/common/tests/acl_util_real
+    run_test build/src/common/tests/acl_principal_tests
+    run_test build/src/common/tests/acl_real_tests
     run_test build/src/iosrv/tests/drpc_progress_tests
     run_test build/src/iosrv/tests/drpc_handler_tests
     run_test build/src/iosrv/tests/drpc_listener_tests
